tokenizer/requirements.txt
accelerate==0.34.0
diffusers==0.30.0
flash_attn==2.6.3
mmengine==0.10.4
onnxruntime==1.20.1
opencv_python==4.10.0.84
pytorch_fid==0.3.0
torch==2.4.0
torchvision==0.19.0
transformers==4.44.0
wandb
tqdm
einops
numpy
omegaconf
Pillow
Requests
safetensors
